Change to INTEGER(11)

+++ Codigo/dindin-api/dindin-app/app/database/migrations/20220309141702-create-user.js
@@ -6,7 +6,7 @@ module.exports = {
     await queryInterface.createTable('user',
       {
         id: {
-          type: DataTypes.INTEGER.UNSIGNED,
+          type: DataTypes.INTEGER(11).UNSIGNED,
           primaryKey: true,
           autoIncrement: true,
           allowNull: false
